Reorganize some social pages constants
